move `checkProjectDependencies` to `config.js`
