+++ lib/utils/config.js
@@ -2,6 +2,7 @@
 require('micro-es7-shim');
 require('coffee-script').register();
 
+const checkDeps = require('check-dependencies');
 const sysPath = require('universal-path');
 const anymatch = require('anymatch');
 const logger = require('loggy');
@@ -445,6 +446,28 @@ const tryToLoad = (configPath, fallbackHandler) => {
   });
 };
 
+const checkProjectDependencies = config => {
+  const packageDir = config.paths.root;
+  const scopeList = config._normalized.isProduction ?
+    ['dependencies'] :
+    ['dependencies', 'devDependencies'];
+
+  return checkDeps({packageDir, scopeList}).then(out => {
+    if (out.depsWereOk) return;
+    const pkgs = out.error.filter(x => x.includes(':')).map(x => x.split(':', 1)[0]);
+    throw pkgs;
+  }).catch(pkgs => {
+    // filter out symlinked packages
+    const pkgPath = pkg => sysPath.join(packageDir, 'node_modules', pkg);
+    const isNotSymlink = pkg => helpers.isSymlink(pkgPath(pkg)).then(x => !x);
+    return helpers.asyncFilter(pkgs, isNotSymlink).then(unmetPkgs => {
+      if (!unmetPkgs.length) return;
+      logger.info(`Using outdated versions of ${unmetPkgs.join(', ')}, trying to update to match package.json versions`);
+      return install({rootPath: packageDir, pkgType: 'package'});
+    });
+  }).then(() => config);
+};
+
 const dontFreeze = ['static', 'overrides'];
 
 const loadConfig = (persistent, options, fromWorker) => {
@@ -460,6 +483,7 @@ const loadConfig = (persistent, options, fromWorker) => {
     .then(config => helpers.deepAssign(config, params))
     .then(trimTrailingSlashes)
     .then(normalizeConfig)
+    .then(checkProjectDependencies)
     .then(fromWorker || addPackageManagers)
     .then(helpers.deepFreeze(dontFreeze));
 };
