+++ lib/watch.js
@@ -6,7 +6,6 @@ const chokidar = require('chokidar');
 const debug = require('debug')('brunch:watch');
 const speed = require('since-app-start');
 const serveBrunch = require('serve-brunch');
-const checkDeps = require('check-dependencies');
 const cpus = require('os').cpus().length;
 const deppack = require('deppack');
 const install = require('deps-install');
@@ -63,28 +62,6 @@ const getWatchedPaths = config => {
   return filterNonExistentPaths(watched);
 };
 
-const checkProjectDependencies = config => {
-  const packageDir = config.paths.root;
-  const scopeList = config._normalized.isProduction ?
-    ['dependencies'] :
-    ['dependencies', 'devDependencies'];
-
-  return checkDeps({packageDir, scopeList}).then(out => {
-    if (out.depsWereOk) return;
-    const pkgs = out.error.filter(x => x.includes(':')).map(x => x.split(':', 1)[0]);
-    throw pkgs;
-  }).catch(pkgs => {
-    // filter out symlinked packages
-    const pkgPath = pkg => sysPath.join(packageDir, 'node_modules', pkg);
-    const isNotSymlink = pkg => helpers.isSymlink(pkgPath(pkg)).then(x => !x);
-    return helpers.asyncFilter(pkgs, isNotSymlink).then(unmetPkgs => {
-      if (!unmetPkgs.length) return;
-      logger.info(`Using outdated versions of ${unmetPkgs.join(', ')}, trying to update to match package.json versions`);
-      return install({rootPath: packageDir, pkgType: 'package'});
-    });
-  });
-};
-
 const setDefaultJobsCount = jobs => {
   const MAX_JOBS = 32;
   const env = process.env.BRUNCH_JOBS;
@@ -126,10 +103,6 @@ class BrunchWatcher {
     application.loadConfig(persistent, options)
       .then(cfg => {
         this.config = cfg;
-        return checkProjectDependencies(cfg)
-          .then(() => cfg, () => cfg);
-      })
-      .then(cfg => {
         if (options.jobs > 1) {
           workers.init(options, cfg);
         }
